feat(json-crdt-peritext-ui): 🎸 use the new <CharOverlay> in debug highlighting
